realtime-pipeline/simulator/requirements.txt
google-cloud-pubsub==2.21.1
